+++ lib/index.ts
@@ -14,7 +14,7 @@
  */
 
 export { PptxRenderer } from './pptx-renderer.js';
-export type { MeasureTextFn, PptxRendererOptions, LogLevel } from './pptx-renderer.js';
+export type { MeasureTextFn, PptxRendererOptions, LogLevel, SlideComment, CommentAuthor } from './pptx-renderer.js';
 export { DEFAULT_FONT_FALLBACKS } from './font-fallbacks.js';
 export type { FontFallbackMap } from './font-fallbacks.js';
 export { bytesToBase64, crc32 } from './utils.js';